requirements.txt
fastapi
uvicorn
motor
groq
langgraph
langchain-core
python-dotenv
pydantic